feat: 여행지 좋아요수 필드 추가

+++ src/main/java/com/ssafy/trip/domain/like/dto/LikeDto.java
@@ -9,5 +9,5 @@ import lombok.*;
 public class LikeDto {
 
     private Long userId;
-    private Long spotId;
+    private int spotId;
 }
